+++ Faker/Generators/ShortGenerator.cs
@@ -9,7 +9,7 @@
 
         object IValueGenerator.Generate(Type typeToGenerate, GeneratorContext context)
         {
-            return context.Random.Next(1, short.MaxValue);
+            return (short)context.Random.Next(1, short.MaxValue);
         }
     }
 }
